pyproject.toml
[build-system]
requires = ["hatchling>=1.18.0"]
build-backend = "hatchling.build"

[project]
name = "legal-rag"
version = "0.1.0"
description = "Graph-aware RAG system for Legal-RAG"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }

authors = [
  { name = "Fan Luo", email = "luofan813@gmail.com" }
]

dependencies = [
  #  Web / API 
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",

  #  数据模型 / 配置 
  "pydantic>=2.0.0",
  "pyyaml>=6.0",

  #  日志 
  "loguru>=0.7",

  #  数值 / 数据处理 
  "numpy",
  "pandas",
  "matplotlib>=3.8.0",
  "scipy>=1.10",

  # Retrieval
  "faiss-cpu>=1.7.4",
  "rank-bm25>=0.2.2",
  "jieba>=0.42.1",

  # Embeddings 
  "sentence-transformers",
  "FlagEmbedding>=1.2.10",

  # ColBERT
  "colbert-ai>=0.2.19", 

  # Providers
  "openai>=1.0",

  #  UI / HTTP 
  "requests>=2.31",
  "httpx>=0.25",

  # Misc
  "tqdm>=4.66",
  "pdfplumber",
  "jedi>=0.16",
  "networkx",
  "pyvis"
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0",
  "pytest-cov>=4.0",
  "black>=24.0",
  "isort>=5.13",
  "mypy>=1.8",
]

notebooks = [
  "jupyter>=1.0",
  "ipykernel>=6.29",
  "notebook>=7.0",
]

pdf = [
  "pypdf>=4.0",
  "pdf2image>=1.17",
  "pytesseract>=0.3.13",
  "Pillow>=10.0",
]

[tool.hatch.build.targets.wheel]
packages = ["legalrag"]

[tool.hatch.build.targets.sdist]
include = [
  "legalrag",
  "scripts",
  "ui",
  "notebooks",
  "data",
  "README.md",
  "LICENSE",
]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra"
testpaths = ["tests"]

[tool.black]
line-length = 88

[tool.isort]
profile = "black"

[tool.mypy]
ignore_missing_imports = true
strict = false
